+++ deployments/operations_test.go
@@ -67,9 +67,9 @@ func testOperationImplementationArtifactPrimary(t *testing.T, deploymentID strin
 			implType, err := GetOperationImplementationType(ctx, deploymentID, "", tt.args.typeName, tt.args.operation)
 			require.NoError(t, err)
 			assert.Equal(t, tt.want.implementationType, implType)
-			_, primary, err := GetOperationPathAndPrimaryImplementation(ctx, deploymentID, "", tt.args.typeName, tt.args.operation)
+			opImpl, err := GetOperationImplementation(ctx, deploymentID, "", tt.args.typeName, tt.args.operation)
 			require.NoError(t, err)
-			assert.Equal(t, tt.want.primary, primary)
+			assert.Equal(t, tt.want.primary, opImpl.Primary)
 		})
 	}