+++ server/server.go
@@ -32,6 +32,7 @@ import (
 	"github.com/ystia/yorc/v4/prov/monitoring"
 	"github.com/ystia/yorc/v4/prov/scheduling/scheduler"
 	"github.com/ystia/yorc/v4/rest"
+	"github.com/ystia/yorc/v4/storage"
 	"github.com/ystia/yorc/v4/tasks/workflow"
 )
 
@@ -68,13 +69,19 @@ func RunServer(configuration config.Configuration, shutdownCh chan struct{}) err
 
 	consulutil.InitConsulPublisher(maxConsulPubRoutines, client.KV())
 
+	// Load main stores used for deployments, logs, events
+	err = storage.LoadStores()
+	if err != nil {
+		return err
+	}
+
 	err = registerBuiltinTOSCATypes()
 
 	if err != nil {
 		return err
 	}
 
-	err = setupConsulDBSchema(client)
+	err = setupConsulDBSchema(configuration, client)
 	if err != nil {
 		return err
 	}