+++ storage/encoding/gob.go
@@ -1,0 +1,42 @@
+// Copyright 2019 Bull S.A.S. Atos Technologies - Bull, Rue Jean Jaures, B.P.68, 78340, Les Clayes-sous-Bois, France.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package encoding
+
+import (
+	"bytes"
+	"encoding/gob"
+)
+
+// GobCodec encodes/decodes Go values to/from gob.
+// You can use encoding.Gob instead of creating an instance of this struct.
+type GobCodec struct{}
+
+// Marshal encodes a Go value to gob.
+func (c GobCodec) Marshal(v interface{}) ([]byte, error) {
+	buffer := new(bytes.Buffer)
+	encoder := gob.NewEncoder(buffer)
+	err := encoder.Encode(v)
+	if err != nil {
+		return nil, err
+	}
+	return buffer.Bytes(), nil
+}
+
+// Unmarshal decodes a gob value into a Go value.
+func (c GobCodec) Unmarshal(data []byte, v interface{}) error {
+	reader := bytes.NewReader(data)
+	decoder := gob.NewDecoder(reader)
+	return decoder.Decode(v)
+}