+++ tosca/entry_schema.go
@@ -16,7 +16,7 @@ package tosca
 
 // An EntrySchema is the representation of a TOSCA Entry Schema
 type EntrySchema struct {
-	Type        string `yaml:"type"`
-	Description string `yaml:"description,omitempty"`
+	Type        string `yaml:"type" json:"type"`
+	Description string `yaml:"description,omitempty" json:"description,omitempty"`
 	//Constraints []ConstraintClause `yaml:"constraints,omitempty"`
 }